+++ source/RTSCamera/src/Patch/Naval/Patch_NavalShipVisualOrderProvider.cs
@@ -40,7 +40,7 @@ namespace RTSCamera.Patch.Naval
 
         public static void Postfix_GetOrders(ref MBReadOnlyList<VisualOrderSet> __result)
         {
-            var newOrder = new NavalToggleShipOrderOrder("order_toggle_ai", GameTexts.FindText("rts_camera_ai_control_ship_on"), GameTexts.FindText("rts_camera_ai_control_ship_off"));
+            var newOrder = new NavalToggleShipOrderOrder("order_toggle_ai", GameTexts.FindText("str_rts_camera_ai_control_ship_on"), GameTexts.FindText("str_rts_camera_ai_control_ship_off"));
             if (Utilities.Utility.ShouldAddToggleShipOrderOrder())
             {
                 if (Input.IsGamepadActive)